new test run with main params
this is a test run of what our main run will probably look like. 25 different runs, 1000 MC simulations, and up to 50 rounds per game.

Need to fix doubling down and split logic, currently the bot is doubling down when it cant afford to do so so it ends up with negative bankrolls